% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{createPOLYspaceObject}
\alias{createPOLYspaceObject}
\title{Create a POLYspace object}
\usage{
createPOLYspaceObject(samples, id_of_samples = NULL)
}
\arguments{
\item{samples}{A list of per-sample tables. Each table contains
#'   columns such as \code{id}, \code{x}, \code{y}, \code{celltype}, and
#'   (optionally) \code{region} if region information is available.}

\item{id_of_samples}{Optional character vector of sample IDs; if \code{NULL},
IDs are generated in order as "1", "2", "3", ...}
}
\value{
A \code{POLYspace} object.
}
\description{
Constructs a \code{POLYspace} container from input samples.
}
