% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_cortex}
\alias{dataset_cortex}
\title{Mouse somatosensory cortex seqFISH+ dataset}
\format{
A list with the following structure:
\describe{
\item{samples}{A named list of samples. Each sample contains a data.table with:
\describe{
\item{id}{Cell identifier.}
\item{x}{x-coordinate of the cell.}
\item{y}{y-coordinate of the cell.}
\item{celltype}{Cell type annotation.}
\item{region}{Region annotation.}
}
}
\item{sample_id}{A character vector of sample identifiers.}
}
}
\source{
Eng, C.-H. L. et al. Transcriptome-scale super-resolved imaging in tissues by rna seqfish+. Nature 568, 235–239 (2019).460
}
\usage{
dataset_cortex
}
\description{
A processed spatial dataset from a single seqFISH+ slide of the mouse somatosensory cortex.
}
\keyword{datasets}
