% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_diabetes}
\alias{dataset_diabetes}
\title{Human type 1 diabetes imaging mass cytometry dataset}
\format{
A list with the following structure:
\describe{
\item{samples}{A named list of samples. Each sample contains a
\code{data.table} with:
\describe{
\item{id}{Cell identifier.}
\item{x}{x-coordinate of the cell.}
\item{y}{y-coordinate of the cell.}
\item{celltype}{Cell type annotation.}
\item{region}{Region annotation.}
}
}
\item{sample_id}{A character vector of sample identifiers.}
\item{group}{A character vector indicating disease-related
stratification.}
\item{subject_id}{A character vector of subject identifiers.}
\item{des}{A list containing a design matrix specifying variables used in regression analyses.}
\item{coef}{A list containing regression coefficients corresponding to variables tested in the model.}
}
}
\source{
Damond, N. et al. A map of human type 1 diabetes progression by imaging mass cytometry. Cell metabolism 29, 755–768526
(2019).527
}
\usage{
dataset_diabetes
}
\description{
A processed spatial dataset from human type 1 diabetes imaging mass cytometry (IMC).
}
\keyword{datasets}
