% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_melanoma}
\alias{dataset_melanoma}
\title{Human melanoma imaging mass cytometry dataset}
\format{
A list with the following structure:
\describe{
\item{samples}{A named list of samples. Each sample contains a
\code{data.table} with:
\describe{
\item{id}{Cell identifier.}
\item{x}{x-coordinate of the cell.}
\item{y}{y-coordinate of the cell.}
\item{celltype}{Cell type annotation.}
\item{region}{Region annotation.}
}
}
\item{sample_id}{A character vector of sample identifiers.}
\item{group}{A factor or character vector indicating immune infiltration degree.}
\item{subject_id}{A character vector of subject identifiers.}
\item{des}{A list containing a design matrix specifying variables used in regression analyses.}
\item{coef}{A list containing regression coefficients corresponding to variables tested in the model.}
}
}
\source{
Hoch, T. et al. Multiplexed imaging mass cytometry of the chemokine milieus in melanoma characterizes features of the477
response to immunotherapy. Sci. immunology 7, eabk1692 (2022).478
}
\usage{
dataset_melanoma
}
\description{
A processed spatial dataset from human melanoma imaging mass cytometry (IMC).
}
\keyword{datasets}
