% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findUniqueComposition.R
\name{findUniqueComposition}
\alias{findUniqueComposition}
\title{Identify neighborhoods with unique compositions}
\usage{
findUniqueComposition(neighborhood, delimiter_node = "--")
}
\arguments{
\item{neighborhood}{A list of neighborhoods.}

\item{delimiter_node}{A string used to join nodes. Default is "--".}
}
\value{
A list of neighborhoods with unique compositions.
}
\description{
Drops duplicates by composition and returns the unique neighborhoods.
}
