% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellNeighborhoodGlyph.R
\name{generateArcData}
\alias{generateArcData}
\title{Generate arc geometry for a neighborhood glyph}
\usage{
generateArcData(
  node_data,
  arc_center,
  arc_radius = c(1, 0.9),
  arc_scale = c(0.7, 1),
  arc_angle = c(3 * pi/10, 7 * pi/10),
  arc_length_out = 100
)
}
\arguments{
\item{node_data}{A data.frame of node-level information, typically
generated by \code{generateNodeData()}.}

\item{arc_center}{A numeric vector of length 2 specifying the (x, y)
center of the glyph.}

\item{arc_radius}{A numeric vector specifying outer and inner arc radii.
Default is \code{c(1, 0.9)}.}

\item{arc_scale}{A numeric vector of length 2 specifying anisotropic
scaling factors applied to arc geometry in the x and y directions.
The first value corresponds to the x-axis scale, and the second
to the y-axis scale. Default is \code{c(0.7, 1)}.}

\item{arc_angle}{A numeric vector of length 2 specifying the start and
end angles (in radians) of the arc. Default is
\code{c(3*pi/10, 7*pi/10)}.}

\item{arc_length_out}{Integer specifying the number of points used
to approximate each arc.}
}
\value{
A list containing arc geometry.
}
\description{
Compute arc-level geometry used to render region- or cell-type–specific
arcs in a cell neighborhood glyph.
}
