% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIsomorphismAgent.R
\name{generateIsomorphismAgent}
\alias{generateIsomorphismAgent}
\title{Generate the isomorphism agent for a target}
\usage{
generateIsomorphismAgent(target, delimiter_node = "--")
}
\arguments{
\item{target}{An adjacency matrix defining the target.}

\item{delimiter_node}{A character string used to join node labels.
Default is "--".}
}
\value{
A data.table containing the isomorphism mappings for
\code{target}.
}
\description{
Builds isomorphism mappings for a target defined by an adjacency matrix.
}
