% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomorphismMapping.R
\name{isomorphismMapping}
\alias{isomorphismMapping}
\title{Get the canonical label for each neighborhood}
\usage{
isomorphismMapping(subs, ta, ag, delimiter_node = "--", noi = NULL, coi = NULL)
}
\arguments{
\item{subs}{Neighborhoods identified by POLYspace.}

\item{ta}{The shape of the neighborhoods.}

\item{ag}{The isomorphism mappings.}

\item{delimiter_node}{A delimiter used when constructing labels. Default is \code{"--"}.}

\item{noi}{Neighborhoods of interest.}

\item{coi}{Compositions of interest.}
}
\value{
Annotations to each neighborhood.
}
\description{
Computes the canonical label for each neighborhood.
}
