% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYspace-package.R
\docType{package}
\name{POLYspace-package}
\alias{POLYspace}
\alias{POLYspace-package}
\title{POLYspace: POLYspace reveals domain-specific cellular neighborhoods through topology-aware spatial analysis}
\description{
Tools to discover and analyze cellular neighborhoods of arbitrary topology accounting for spatial domains.
}
\author{
\strong{Maintainer}: Huimin Wang \email{huimin.wang@duke.edu}

}
\keyword{internal}
