% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveSingleResult.R
\name{retrieveSingleResult}
\alias{retrieveSingleResult}
\title{Retrieve single-sample enrichment results}
\usage{
retrieveSingleResult(
  object,
  targets = "default",
  neighborhoods_of_interest = c("union", "intersect")
)
}
\arguments{
\item{object}{A \code{POLYspace} object or a list of \code{POLYspace} objects.}

\item{targets}{Targets to include.}

\item{neighborhoods_of_interest}{Character vector (e.g., \code{c("union","intersect")}),
or a user-specified set of neighborhoods to include.}
}
\value{
A list with components \code{results} (data.table of
single-sample enrichment) and \code{targets} (the targets used).
}
\description{
Extracts per-sample enrichment results from a \code{POLYspace} object or from
a list of \code{POLYspace} objects.
}
