% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{showWelcomeMessage}
\alias{showWelcomeMessage}
\title{Show a welcome message}
\usage{
showWelcomeMessage(POLYspace)
}
\arguments{
\item{POLYspace}{A \code{POLYspace} object.}
}
\value{
No return value, called for side effects.
}
\description{
Prints a brief welcome/summary for a POLYspace object.
}
