% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{toComposition}
\alias{toComposition}
\title{Convert segments to corresponding compositions}
\usage{
toComposition(segments, delimiter_node = "--")
}
\arguments{
\item{segments}{Segment labels.}

\item{delimiter_node}{Delimiter used to join labels (default \code{"--"}).}
}
\value{
A string of corresponding compositions.
}
\description{
Collapses segment labels into a composition string.
}
