% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodMining.R
\name{neighborhoodMining}
\alias{neighborhoodMining}
\title{Mine neighborhoods}
\usage{
neighborhoodMining(
  POLYspace,
  targets = list("singlet", "pair", matrix(c(0, 1, 0, 1, 0, 1, 0, 1, 0), 3, 3)),
  ncores = 1,
  mc.preschedule = TRUE
)
}
\arguments{
\item{POLYspace}{A \code{POLYspace} object.}

\item{targets}{A list specifying neighborhood shapes of interest, such as
\code{"singlet"}, \code{"pair"}, or a custom adjacency matrix.
By default, targets include singlet, pair, and triple-chain neighborhoods.}

\item{ncores}{Integer; number of CPU cores to use. This is particularly useful
when targets are more complex than singlet or pair and when multiple samples
are included in the \code{POLYspace} object. In this case, computations are
distributed across cores by combinations of samples and targets}

\item{mc.preschedule}{Logical; passed to \code{parallel::mclapply()}.}
}
\value{
An updated \code{POLYspace} object whose \code{neighborhoods} slot includes
the discovered neighborhoods. Each element of the list corresponds to a sample;
within each sample, each sub-list element corresponds to a target, with ordering
consistent with the \code{targets} slot.
}
\description{
Finds specified neighborhoods in a \code{POLYspace} object.
}
