% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFeatureMatrix.R
\name{retrieveFeatureMatrix}
\alias{retrieveFeatureMatrix}
\title{Retrieve a neighborhood feature matrix}
\usage{
retrieveFeatureMatrix(
  object,
  region = TRUE,
  targets = "default",
  neighborhoods_of_interest = c("union", "intersect")
)
}
\arguments{
\item{object}{A \code{POLYspace} object or a list of \code{POLYspace} objects.}

\item{region}{Logical; use region-level neighborhoods.}

\item{targets}{Targets to include.}

\item{neighborhoods_of_interest}{Character vector (e.g., \code{c("union","intersect")}),
or a user-specified set of neighborhoods of interest.}
}
\value{
A list containing the feature matrix, neighborhood information, and targets.
}
\description{
Builds a feature matrix from a \code{POLYspace} object or a list of
\code{POLYspace} objects.
}
